fix: refacto the samples examples and core
