+++ python/examples/youtube_translation.py
@@ -21,7 +21,7 @@ transcription = gladia_client.transcribe(
                     "value": "Misha",
                     "pronunciations": ["micha, misha, mi cha, mi sha"],
                     "intensity": 0.4,
-                    "language": "de",
+                    "language": "ko",
                 },
             ],
             "default_intensity": 0.6,
