+++ typescript/core-concepts/pre-recorded/pre-recorded.ts
