+++ typescript/interface.d.ts
